EGL Multifile Blobcache: Update size on key reuse
Before this fix, if a key was submitted with a different value, we
would increase the size of the cache by the new amount. Some
platforms are reusing keys a lot and running out of memory.

Now we check to see if it is contained, then remove it before adding
the new value.

This also contains an optimization to removeFromHotCache. Previously
this would wait until all pending writes had completed. This was okay
since we only removed when caches were overflowing. With key reuse,
we are removing entries more often. We only really need to wait if
there are pending writes for the entry being removed.

Test: libEGL_test, EGL_test, ANGLE trace tests, apps
Bug: b/351867582
Change-Id: I66fe9cde18e468e541a80296c80f2234ac61acb0
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>